requirements.txt
absl-py==2.1.0
accelerate==0.24.0
antlr4-python3-runtime==4.9.3
asciitree==0.3.3
atomics==1.0.2
av==13.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.0.0
contourpy==1.3.0
cycler==0.12.1
diffusers==0.18.2
dill==0.3.8
dm-env==1.6
dm-tree==0.1.8
dm_control==1.0.23
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
etils==1.9.4
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
gymnasium==1.0.0
h5py==3.12.1
huggingface-hub==0.25.2
hydra-core==1.3.2
idna==3.10
imagecodecs==2024.6.1
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
labmaze==1.0.6
llvmlite==0.40.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
mujoco==3.2.4
networkx==3.3
numba==0.57.1
numcodecs==0.12.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.28.2
psutil==6.0.0
pycparser==2.22
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.1.4
pyrealsense2==2.55.1.6486
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
scipy==1.14.1
sentry-sdk==2.16.0
setproctitle==1.3.3
Shimmy==2.0.0
shtab==1.7.1
six==1.16.0
smmap==5.0.1
spnav==0.9
sympy==1.13.3
termcolor==2.5.0
threadpoolctl==3.5.0
timm==0.9.7
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.4
transforms3d==0.4.2
triton==3.0.0
typing_extensions==4.12.2
tyro==0.8.5
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.3
zarr==2.16.1
zipp==3.20.2